requirements.txt
Django==5.0
django-sesame==3.2
